Multi-Agent-Customer-Support/requirements.txt
gradio>=5.29.0
langchain>=1.0.0
langchain-ollama>=1.0.0
langchain-community>=0.4.0
langchain-core>=1.0.0
langgraph>=1.0.0
langgraph-checkpoint>=4.0.0
langgraph-prebuilt>=1.0.0
langgraph-supervisor>=0.0.20
sqlalchemy>=2.0.0
pydantic>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0

